Make weather text reporting views generic
